refactor: ChangeViewButtonStackView 분리
