+++ Expo1900/Expo1900/Model/ExpositionItemEntity.swift
@@ -11,7 +11,7 @@ struct ExpositionItemEntity: Decodable, Equatable {
     let shortDescription: String
     let description: String
     
-    enum CodingKeys: String, CodingKey {
+    private enum CodingKeys: String, CodingKey {
         case name
         case imageName = "image_name"
         case shortDescription = "short_desc"
